+++ src/components/calculator/Calculator.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import calculate from './logic/calculate';
+import calculate from '../logic/calculate';
 import './Calculator.scss'
 function Calculator() {
   const [result, setResult] = useState({
