+++ src/components/calculator/Calculator.scss
