+++ src/components/common/CloseButton.scss
