+++ src/components/common/Navbar.scss
@@ -9,7 +9,7 @@
   border-bottom: 2px solid rgb(139, 139, 139);
   height: 100px;
   background-color: rgb(31, 22, 71);
-  background-image: url('./dan-cristian-padure-h3kuhYUCE9A-unsplash.jpg');
+  background-image: url('../img/dan-cristian-padure-h3kuhYUCE9A-unsplash.jpg');
   background-repeat: no-repeat;
   background-size: cover;
 }
@@ -36,7 +36,7 @@
   left: 0;
   height: 100%;
   width: 100%;
-  background-image: url('./dan-cristian-padure-h3kuhYUCE9A-unsplash.jpg');
+  background-image: url('../img/dan-cristian-padure-h3kuhYUCE9A-unsplash.jpg');
   background-repeat: no-repeat;
   background-size: cover;
   font-size: 30px;
@@ -62,12 +62,23 @@
     }
   }
 
+  .container-nav {
+    height: 120px;
+  }
+
+  .navbar-title {
+    padding: 40px;
+    font-size: 40px;
+  }
+
   .navbar-nav {
     display: flex;
     flex-direction: row;
-    gap: 20px;
+    gap: 30px;
     list-style: none;
     padding: 10px;
+    margin-right: 40px;
+    margin-top: 40px;
 
     .navbar-link {
       display: flex;
