+++ src/components/home/Home.scss
@@ -12,7 +12,7 @@
 .problem-container {
   width: 300px;
   height: 250px;
-  background-image: url('./pizarra-vacia.jpg');
+  background-image: url('../img/pizarra-vacia.jpg');
   background-repeat: no-repeat;
   background-size: cover;
   color: white;
