+++ src/components/quote/Quote.js
