+++ src/components/quote/Quote.scss
